+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "facilities/base"]
+	path = facilities/base
+	url = git@github.com:bitfinexcom/bfx-facs-base.git
